chore: Move slice-into-uninit regression test
Move the test to integration tests since it only exercises public API.
Modify the test to ensure that the size of DerefMut output
for Slice is as expected and writing into the slice works.
